+++ src/manager/collectionManager/CollectionComponent.vue
@@ -114,6 +114,10 @@
                 props: {
                   windowId,
                 },
+                state: {
+                  headerTitle: 'collectionManager.more',
+                  headerIcon: '$vcsComponents',
+                },
                 provides: {
                   collectionComponent,
                 },