+++ src/manager/collectionManager/CollectionComponentList.vue
@@ -1,20 +1,35 @@
 <template>
+  <div>
+    <div class="v-expansion-panel-header px-2 v-expansion-panel-header--active">
+      <div class="d-flex justify-space-between">
+        <div class="d-flex align-center">
+          {{ $t(title) }}
+        </div>
+        <VcsActionButtonList
+          v-if="actions?.length > 0"
+          :actions="actions"
+          :overflow-count="overflowCount"
+        />
+      </div>
+    </div>
   <vcs-list
     :items="items"
     :draggable="draggable"
     :selectable="selectable"
     :single-select="singleSelect"
     v-model="selection"
-    :title="title"
+      :show-title="false"
     @item-moved="move"
     @item-renamed="rename"
   />
+  </div>
 </template>
 
 <script>
   import { inject } from 'vue';
   import { IndexedCollection } from '@vcmap/core';
   import VcsList from '../../components/lists/VcsList.vue';
+  import VcsActionButtonList from '../../components/buttons/VcsActionButtonList.vue';
 
   /**
    * Moves an item to a new position.
@@ -64,6 +79,7 @@
   export default {
     name: 'CollectionComponentList',
     components: {
+      VcsActionButtonList,
       VcsList,
     },
     props: {
@@ -85,6 +101,9 @@
         draggable: collectionComponent.draggable,
         selectable: collectionComponent.selectable,
         singleSelect: collectionComponent.singleSelect,
+        overflowCount: collectionComponent.overflowCount,
+        limit: collectionComponent.limit,
+        actions: collectionComponent.getActions(),
         move({ item, targetIndex }) {
           moveItem(collectionComponent, item, targetIndex);
         },